==> markdown/02_inventory_root_directory.Rmd <==
---
title: "Inventory the current file directory"
date: "Created: 2018-01-03 <br> Updated: `r Sys.Date()`"
output: 
  html_notebook:
    toc: true
    toc_float: true
    css: custom-css.css
---

# Overview {#overview}

I needed to come up with a better method for tracking notes/documentation/file dependencies in analysis projects. The goal of Pathfinder is to help track file dependencies in research projects created in RStudio. Perhaps other things later. 

Assumptions:    
1. Pathfinder will typically be deployed in R projects.   

  * When deployed in R projects, Pathfinder will automatically detect the .Rproj file and the path to that .Rproj file will be set as the root directory.   

  * If the user desires a different behavior (e.g., target a different project than the one the user is currently working in or mapping dependencies among a collection of files that do not exist in an R project) then the path can be specified manually.    

    - The specified path will serve as the root directory. Pathfinder will attempt to gather dependency information about every file in the root directory and all child directories.   
        
    - If the user is not in an R project and a path is not specified, then Pathfinder will set the current directory to be the project root. Pathfinder will attempt to gather dependency information about every file in the root directory and all child directories.
        
2. Therefore, to use Pathfinder the user must be in an open R project, provide a file path, or both.


# Load packages

```{r setup, include=FALSE}
knitr::opts_chunk$set(comment = NA)
Sys.setenv(TZ = "US/Central")
```

```{r message=FALSE}
library(tidyverse)
source("../r/get_project_root.R")
source("../r/check_for_rproj.R")
```










# Find root and create a list of all files under the root {#root}

## Save the path to the current project root directory

Make here::here() the default location for finding the project root, but allow the user to override it. That should also allow me to separate the package itself from the package development/testing files. 

### Testing default behavior - inside an R project

```{r}
get_project_root()
```

### Testing user-provided path

```{r}
get_project_root("/Users/bradcannell/Dropbox/R/Packages/pathfinder_example_project")
```

```{r}
root_dir <- get_project_root("/Users/bradcannell/Dropbox/R/Packages/pathfinder_example_project")
root_dir
```


## Create a vector containing the names of all the files in the root directory

```{r}
root_dir_file_list <- list.files(path = root_dir, recursive = TRUE)
root_dir_file_list
```


## Test to make sure there is an .Rproj file

```{r}
check_for_rproj(file_list = root_dir_file_list)
```

[top](#top)










# Save root path and root directory file list for future {#save-root}

```{r}
readr::write_rds(root_dir, "../data/root_dir.rds")
readr::write_rds(root_dir_file_list, "../data/root_dir_file_list")
```

[top](#top)

&nbsp;

-------------------------------------------------------------------------------

```{r echo=FALSE}
sessionInfo()
```
